PsyPhyOculusDemo
Created the demo program PsyPhyOculusDemo.
Had to perform a lot of structural changes to make OpenGLObjects and
VectorsMixin compatible with OVR, including the creation of the PsyPhy
namespace.

To get OpenGLObjects to render properly using the window created as in
the Oculus demo programs I had to inhibit backface culling.