Enable tracking of vcxproj.user.
Up until now. *.user is ignored by Git. But *.vxproj.user includes the
Debugging->Command Arguments and Debugging->Working Directory
properties, so I re-enable *.user in .gitignore.  It may be necessary to
ignore other, more specific .user files, but so far I don't see any. It
might also be annoying to share these setting across users, but I think
that for the moment it is advantageous.