Some other little progress
I'm trying to add the moving projectile. I could easely do it in ùy
codes, but with your code structure I having a lot of problem. As soon
as I try to do something slightly more complicated, I'm not able to make
it work because of several issues among with 'variable visibility'. I
think that in order to be able to do something I should  have more
information about how all differents pieces of code are linked or in any
case I would need to know how the whole thing is structured in your
mind.