Correct some comments and describe Git strategy.
If you want to change something, go ahead  and change it, rather than
commenting out the change. The modification will be visible in the
Changes part of GitHub. Then, when you commit the changes, add your
reasons here in the comments to the commit as to why you made the
change. I will see the changes there and can, if need be, revert them.

Try to keep the comments accurate as well. If you copy code from one
place to reuse it elsewhere, be sure to modify the comments
appropriately.