private versus protected for GraspGLObjects class.
When I orginally created GraspGLObjects, it made sense that the
initialization constants be restricted to the class and not accessible
externally. But now GraspGLObjects is inherited to create GraspVR, which
may like to access these same constants. So they are now declared to be
'protected' which means that they can be accessed inside methods that
are part of the inheriting class.