Enable()/Disable() versus ->enabled
Arguably, it is better to call the methods rather than accessing the
property directly, because it allows the object to implement additional
actions when it is enabled or disabled. Indeed, it might be better to
make the 'enabled' property a protected property.

I started out the wrong way without thinking. Here I am reverting to use
the best practice of calling the Enable() and Disable() methods.