Tools to create instruction screens.
The GRASP GUI interface uses an webbrowser window to display text and
images to the subject. Here I provide the means to generate HTML files
to be displayed and a small app to preview the contents. I use pandoc to
create HTML from markdown, with some added HTML code embedded in the .md
file to center the title and the image.