Pose and TrackerPose calls by reference.
This was a big operation to change calls accepting a Pose as a parameter
to be call by reference instead of call by pointer. At the same time I
added functions to handle simple Poses and a TrackerPose now contains a
Pose as an element. This means that to get to the position and
orientation of a TrackerPose one has two indirections to do:
xxx.pose.position and xxx.pose.orientation.