A way to group several objects at the same pose.
Unlike the Assembly object, a yoke is used simply to drive multiple
objects to the same position and orientation. You can also change their
color as a group. As opposed to Assembly, however, the position of a
component is relative to the world, not relative to the assembly and
whenever the yoke is told to change its position or orientation, the
same operation is performed on each component.