Tools and Hands and Colors
Integration of a number of modifications:

There are now three 'tools', one for full visual feedback and two for
kinesthetic feedback. Of the latter, one is used during the target phase
for -K-K and the other for the response phase in V-K, K-K.  The former
varies in color according to the orientation of the hand wrt a reference
roll orientation.  Each has a laser pointer attached to it. I keep them
separate to allow easier modification of one versus another.

The three tools are now grouped together by a new object called a yoke.
It is similar to an Assembly but maintains the position of each object
separately.

The roll angle of the hand can be adjusted with the mouse. One can also
use the arrow keys to adjust the simulated pitch and roll of the hand.
These should be used to bring the laser pointer to the center target.

Color variations were not working and because I did not understand the
equations I decided to try a different approach that makes the color
vary from red to green via yellow. But the sensitivity has to be
adjusted.