Trackers are now a separate entity from GraspVR.
Instead of including the trackers as instance variables in the GraspVR
class, a set of trackers is now created as a separate object and GraspVR
receives a pointer to that set. This makes it easier to switch in and
out different sets of trackers.

At this point, only the Oculus-only set of trackers has been tested.