Scale to mm in OculusMapper
I was poking around trying to find ways to reduce the time in BlastIt()
so as to maintain a higher frame rate and avoid (I hope) glitches. I
found that viewScaleDesc includes a paramter to map the world
coordintates to meters.  I think that this has something to do with time
warping or it may have to do with the depth planes. I set it to 0.001
but I haven't noticed any difference.

There is also the option ovrLayerType_Direct which should turn off
distortion mapping and timewarp in BastIt() as well. I'm still playing
with this stuff.