Close stream and disconnect when shutting down CodaRTnetTracker.
Shutting down and restartingCodaRTnetTracker was crashing because, I
believe, it was leaving the stream connection open and then trying to
reconnect on the same port.

I could make it even more robust by having it cycle through other ports
if the first connection at 7000 fails, but I will leave that to later.