Setting CODA alignment transorm (in progress)
We need to define the alignment transforms based on methods other than
the CODA-provided routine, but the CodaRTnet interface does not provide
this funcitonality. So I am recreating the functionality by rewriting
the file on the CODA server that hold the transform information, using
WinRCP / FTP.

This is far from done.