Reduce ovrSubmitFrame() duration by cropping FOV.
The time required to submit a frame to the Oculus is too long using the
Display Port adaptors that we have, leaving less than 2 ms to do all of
the rendering. The time to submit the frame depends on the size of the
bitmap sent  to the Oculus. By cropping the bitmap on the edges we are
able to reduce the time required to submit the frame, at the expense of
a reduce FOV. This is hopefully just a temporary fix and that we will be
able to restore the full FOV by twiddling parameters elsewhere.