Circular 'glasses'
When the FOV is cropped to decrease the time to submit a frame to the
Oculus, one sees a black rectangular frame around the visual scence. To
counteract the negative effects that might have by providing a reference
frame I have created a new set of 'glasses' that have a rim that changes
color with head orientation but that is black outside the circular rim.