Create GraspTaskManagers
GraspTaskManagers are built on top of GraspVR. They will handle the
progression through a block of trials. The plan is to have a different
TaskManager for each type of task (V-V, K-K,, V-K). Right now I have
only started the base class that does a V-V training, and even that is
not done.