Alignment and Rigid Body models (did not work???)
This combination of rigid body files and alignment did not work. Need to
figure out why. One HMD file was acquired by executing
CreateRigidBodyModelFile.exe inside the debugger. I then executed the
batch file to do all three from the root directory. When that did not
work, we redid the alignment and did the model files again. But that
still did not work. There is a difference between the file done from the
debugger and the other three but that might reflect the change to the
alignment.

I need to figure this out!!!!!!!!!!