Continuing work on rigidbody-based alignment.
All seems to be working on the SM at Tecnalia.
It is easy to forget the serial number issue, though, and work with a
non-aligned system without knowing it. Need to find a way to at least
signal the problem, or better yet, detect the serial numbers and write
out the alignment file accordinly.