Flattened and merged instructions.
Most of the instructions are the same between paradigms (straighten
head, tilt head, feedback) and the two that are specific are shared
between two protocols (V target between V-V and V-K, K response between
V-K and K-K). So I created a common set with file names that make it
easier to know what is in each one. I also grouped all of the images
into a single “Pictures” subdirectory to avoid problems with long paths
to files.