Cleaned up Build configurations.
Ingore default libraries as needed when defaults libraries are mixed.
This happened most often when linking with the CODA SDK.

Used /ignore:4075 to avoid warning. Had to patch link.exe as well to
supress the warning.

Created the Draft Configuration.  Draft builds the projects in Release
mode but does not perform ReseaseGrasp, so no tar file is generated.
Default builds the entire system and places it in a tar respository for
installation on another machine.