Tilt arrow and success based on time not cycles;
Instead of using a cycle counter, I use timers to create hysteresis in
the control of the circular arrow that shows which way to tilt the head.
This should make it more robust.

In addtion, I changed the strategy. As soon as the head is in the good
zone, the arrow goes off, but if one has to be out of the zone for some
time (secondsToBeBad) before the arrow comes back on again. This gives
the subject the chance to correct based just on the color before being
distracted by the arrow.