Debug, TestRelease, DraftRelease, Release
Created 4 compile configurations:

Debug - Used during debugging.
TestRelease - Compile in release mode, but do not create a release
archive.
DraftRelease - Compile in release mode and generate a release archive,
but the archive is labelled 'Draft' to indicate that it is not meant to
be definitive.
Release - This compiles in Release mode and generates an archive that in
theory could be used as an official release.