Fixed HEADER macro and harmonized instruction screens.
The HEADER macro worked when generating an HTML document for display in
GraspGUI, but when the HTML was displayed in a standard web browser or
converted to pdf it either centered the text or magnified it or vice
versa, but not both. I fixed the macro to combine the two and touched up
each of the instruction screen.