Recursive GraspScreenshots
A program to take snapshots of the Grasp VR world in different states. I
have not figured out how to save a window multiple times within the same
program, so I modified the program to do only a single capture of the
window, but then to call itself recursively if there are more than one
command line arguments.