GraspDocumentation and collateral damage
Created a program to produce screenshots from within GRASP and then
created makefiles to generate some simple documentation automatically.
The advantage is that if some of the VR images change or the
instructions then these documents are updated automatically without a
bunch of cut and paste. Still, one needs to be careful that the
makefiles are up to date.