<SPACE> before ; in Align command line definition
GraspGUI tacks on the filename root at the end of the command line. If
the command does something like --file=, it is essentially that there be
no space between the = and the ;.  But if the command simply takes the
filename as the last argument, if there is no space before the ; in the
.tsk file then the filename gets appended to the last argument in the
command line defintion. This should be made to be more robust.