Set a CODA alignment that is suitable for OpenGL
This is a simple transformation that assumes that the CODA units are in
front of the subject and horizontal. The alignment transformation simply
maps CODA intrinsic coordinates to OpenGL coordinates. Note that if you
have multiple codas they will not be aligned to each other. Each one
gives OpenGL coordinates with respect to itself.

I created this for my own convenience. Use it at your own risk.