Read CodaRTnet constants from a .ini file.
This will make it easier to configure for one DEX/Coda system to
another.
CodaRTnet.ini is ignored by GitHub so that you can keep on your machine
the version that corresponds to your hardware. If you need to change
your configuration, edit the qualified version (.GM1, .SM, etc.) then
copy it to CodaRTnet.ini on your machine.