Align to HMD during blanking.
An alignment is set that puts the center of the world at the HMD,
aligned with the subject's gaze. This alignment is applied to tracker
data before moving VR objects to that location.

Note that this is different from boresighting in that all 3 trackers
(HMD, Chest, Hand) undergo the same transformation meaning that after
the alignment the hand will not be at 0 but rather at it's position
relative to the HMD in the new VR reference frame.

I think that this is working based on the simulated trackers in V-VK,
but I have not yet tried it with the CODA.