Indifferent to head roll during response feedback.
In the previoius version we set the desired head roll to zero as soon as
the subject pressed a button to respond so that they can start to
reorient the head already. But this generated a jump to a red halo when
the button got pushed, which could be perceived as an error. What I do
now instead is to simply ignore the head roll (no halo feedback) during
the feedback phase. The halo is presumably already cyan, otherwise an
error would have occurred, and it stays cyan whether the subject starts
to reorient the head towards upright or not.