Workaround for RTnet client memory leak.
The CodaRTnet interface incorrectly reallocates memory each time it
fills a packet, without freeing previously allocated buffers. If one
reuses a RTNetworkPacket or PacketDecode* object, memory use increases
without bound. The solution is to create and destroy the objects within
the loop so that the destructor gets called on each cycle. This in
implemented here by declaring the pacet and decoder objects as dynamic
variables withing the acquisition loop.