Created tilted chest marker model.
This is used to compensate for the tilt of the backrest of the DEX chair
so that the resulting reference frame is aligned with gravity and the
cahir.

Note that we are not 100% sure about the direction of the tilt. Model
files with tilt of +/- 10° of pitch can be found in TweakedChestModels.