GraspDisplays - Work in progress
Create a new module to allow for different display modes for Grasp.exe.
Previously, the Oculus was deeply entwined within the GraspVR module.
This functionality has be abstracted to a new module that for the moment
does exactly the same thing. But in the future it will be possible to
swap different GraspDisplay objects in to switch, for example, between
an Oculus and a desktop version of Grasp.