GraspWindowsDisplay to run on desktop (in progress)
Created a GraspWindowsDisplay object to allow Grasp.exe to be run on a
laptop without an Oculus. Included in this is the creation of
GraspSimulatedTrackers vs. GraspOculusOnlyTrackers.

There is still a lot to be done to get the descktop version to work
properly, notably to handle keyboard inputs and to simulate the
trackers.