Stack overflow bug fix.
The creation of the tracker to use the legacy CodaSDK (for use with the
NASA machine) introduced a bug into Grasp.exe. I moved the data arrays
to the Tracker base classs, but this class was being used as a dynamic
variable in DexServices just to get the CopyMarkerFrame routine. So the
stack was being overrun. I changed the object to a static variable in
DexServices::AddDataSlice(), which fixed the problem.