Cleaned up Apply for OculusViewpoint
A generic Viewpoint need a reference to which window to use, so Apply()
takes a pointer to an OpenGLWindow as an argument. An OculusViewpoint
implicitly goes to the Oculus, so I originally specified its Apply()
method without this parameter. But then it cannot be referred to as a
generic Viewpoint. So I made it compliant with the base class.