Manual trial abort.
By pressing Return on the keyboard or on the remote one can abort a
trial and have it repeated at the end of the block.

This can be used to abort a trial in progress, but does not allow to
signal a trial to be repeated if the validate button was pressed by
mistake.