Set the initial hand position to be out of view.
Initial position of the virtual hand was 0,0,0 as it is by default for
most objects. This means that if the tracker does not see the hand right
away, the virtual hand is at the viewpoint during the initial phases of
the trial. We now  set it too a place below the viewpoint so that by
default it is not in the field of view. However, if the hand is in view,
it will be displayed there.