Don't ignore GraspSequences
I was trying to ignore it in the root directoy because they are
generated each build. But the effect was to ignore the directory in the
Source Code tree because it had the same name. Now the directory in the
root runtime tree is called Sequences so we can ignore it and still
track GraspSequences in the source code tree.