Tracking and installation of documentation files.
Documentation files are no longer maintained in the Source Code
directory. They exist in the GrasMMIDocumentation and
GripMMIDocumentation directories under GripGrasp and they are tracked by
Git.

When creating a runtime set of code, only the installation instructions
for the runtime environment and the release notes are included in the
tar file that is to be transferred to the runtime machine.