Improve end of subsession messages.
First, execute "Grasp.exe --doff", rather than through Grasp.bat, to
avoid transferring files to GRIP at the end.

When the task is first selected, say already that the subject should
move on to the next subsessions or quit according to the PODF and
execution notes. If the step gets executed (e.g. if the subject presses
return on the remote to automatically advance to the next block), show
the same message on exit.