Tweak and Tilt Rigid Body Files
Created a tool to tweak the body files so that the reference frame is
aligned with the back 4 markers.

Then added the possibility to rotate the model in pitch. This is used to
create a model of the chest markers with a 10° upward tilt, to be used
when aligning the codas in the upright position, because the backrest
has a 10° tilt wrt the vertical.