.bat are named .bat.src
This is a hack to get around a limitation of git. If I want to ignore
changes to RunGRASP.bat in the root directory, the entry in .gitignore
causes the version in the source file aborescence to be ignored as well.
So I give it this funny name and change the name when it is copied from
the source files to the runtime root directory.