Transmit full history to ground
Keeps track of when tasks are started and completed, even when stopping
and starting. Then when GraspGUI is idling waiting to start a new task
it cycles through the event list and send it to ground. So even if we
went LOS, the starting and ending times of tasks, and the completion
codes are stored and gradually sent to ground.