Pass a marker frame to a CodaPoseTracker
Instead of changing the assigned marker buffer all of the time, you can now just ask the tracker to compute the pose from a frame that you pass as a parameter.

But you still need to assign a buffer if you want to use the normal GetCurrentPose and GetCurrentPoseIntrinsic functionality of a tracker in real time.