Clean up VectorsMixin
Added functions to read vectors and quaternions from a FILE.

Default formating for output strings is now a state of each instance.

Created a global instance to expose the functionality to functions and classes that cannot inherit VectorsMixin. But it is probably better to make local instances of VectorsMixin as needed.